@@ -9,7 +9,7 @@ public:
 	void send_announce(data_ref);
 	void send_connect(data_ref);
 	void send_scrape(data_ref);
-	void send_error(data_ref, const std::string& msg);
+	void send_error(data_ref, std::string_view msg);
 	Ctransaction(const Csocket&);
 private:
 	const Csocket& m_s;
